requirements.txt
numpy
scipy
scikit-learn
gensim
matplotlib
pandas
datasketch
faiss-cpu    
hnswlib
tqdm
